pyproject.toml
[tool.poetry]
name = "rust_compiler_tests"
version = "0.1.0"
description = ""
authors = ["Tim Jentzsch <tim-jentzsch@gmx.de>"]

[tool.poetry.dependencies]
python = "^3.9"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
